% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_arguments}
\alias{check_arguments}
\title{Check if arguments are valid (internal)}
\usage{
check_arguments(x, ref)
}
\arguments{
\item{x}{\code{list}. The user-query list to compare against \code{ref}.}

\item{ref}{\code{list}. The reference list to compare \code{x} against.}
}
\value{
Error messages or \code{TRUE} if all arguments are valid.
}
\description{
This function checks whether user arguments are valid
in terms of being an available argument or of the correct
class.
}
