% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_function}
\alias{create_function}
\title{Create functions (internal function)}
\usage{
create_function(
  name,
  args = NULL,
  endpoint = NULL,
  fields = c("author", "details", "examples", "export"),
  overwrite = FALSE
)
}
\arguments{
\item{name}{\code{character}. The name of the function to be created (this
will also be the file name).}

\item{args}{\code{character}. The names of the arguments to include in the
function.}

\item{endpoint}{\code{character}. The API endpoint to fetch arguments from
to include in the function.}

\item{fields}{\code{character}. The fields to include in the roxygen2
documentation.}

\item{overwrite}{\code{logical}. Should pre-existing files of the same name
be overwritten? Defaults to \code{FALSE} (strongly advised).}
}
\value{
Nothing. Writes files of functions with desired arguments and
roxygen2 fields as a side effect.
}
\description{
This function can be used to automatically setup functions for Macrostrat
endpoints.
}
\details{
Files are generated in the \verb{./R/} directory using the provided
\code{name} as the file name.
}
