% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_econs.R
\name{def_econs}
\alias{def_econs}
\title{Retrieve economic definitions}
\usage{
def_econs(econ = NULL, econ_type = NULL, econ_class = NULL, econ_id = NULL)
}
\arguments{
\item{econ}{\code{character}. The name of the economic resource (e.g., "gas
reservoir") to return a definition for.}

\item{econ_type}{\code{character}. The type of economic resource (e.g.,
"hydrocarbon") to return a definition for.}

\item{econ_class}{\code{character}. The class of economic resource (e.g.,
"energy") to return a definition for.}

\item{econ_id}{\code{integer}. The unique identification number(s) of the
economic resource(s) to return a definition for.}
}
\value{
A \code{dataframe} containing:
\describe{
\item{econ_id}{The unique identifier of the economic resource}
\item{name}{The name of the economic resource}
\item{type}{The type of economic resource}
\item{class}{The class of economic resource}
\item{color}{The recommended coloring for units based on the dominant
lithology}
\item{t_units}{The total number of Macrostrat units containing the economic
resource}
}
}
\description{
A function to retrieve the definitions for one or more
economic resources in the Macrostrat database. By default, all economic
definitions are returned.
}
\examples{
\dontrun{
if (interactive()) {
  # econ_inf <- def_econs()
  # econ_inf <- def_econs(econ_type = "hydrocarbon")
}
}
}
\seealso{
Geologic features and attributes:
\code{\link{def_environments}()},
\code{\link{def_grain_sizes}()},
\code{\link{def_lithologies}()},
\code{\link{def_lithology_att}()},
\code{\link{def_measurements}()},
\code{\link{def_minerals}()},
\code{\link{def_structures}()}
}
\author{
Bethany Allen
}
\concept{defs_feat}
