% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_lithologies.R
\name{def_lithologies}
\alias{def_lithologies}
\title{Retrieve lithology definitions}
\usage{
def_lithologies(
  lithology = NULL,
  lithology_group = NULL,
  lithology_class = NULL,
  lithology_type = NULL,
  lithology_id = NULL
)
}
\arguments{
\item{lithology}{\code{character}. The name of a lithology (e.g.,
"limestone") to return a definition for.}

\item{lithology_group}{\code{character}. The name of a lithological group
(e.g., "unconsolidated") to return a definition for.}

\item{lithology_class}{\code{character}. The name of a lithological class
(e.g., "sedimentary") to return a definition for.}

\item{lithology_type}{\code{character}. The name of a lithological type
(e.g., "siliciclastic") to return a definition for.}

\item{lithology_id}{\code{integer}. The unique identification number(s) of
one or more lithologies to return a definition for.}
}
\value{
A \code{data.frame} containing the following columns:
\itemize{
\item \code{lith_id}: Unique ID of the lithology.
\item \code{name}: The name of the lithology.
\item \code{type}: Lithology type, less inclusive than class.
\item \code{group}: Lithology group, less inclusive than type.
\item \code{class}: Lithology class, more inclusive than type.
\item \code{color}: Recommended coloring for lithology.
\item \code{fill}: The code for the fill pattern as established in the
\href{https://ngmdb.usgs.gov/fgdc_gds/}{Federal Geographic Data Committee}'s
\href{https://ngmdb.usgs.gov/fgdc_gds/}{Digital Cartographic Standard for Geologic Map Symbolization}.
\item \code{t_units}: The total number of Macrostrat units that are
partially or entirely composed of the lithology.
}
}
\description{
A function to retrieve the definitions for one or more
lithologies in the Macrostrat database. By default, all lithology
definitions are returned.
}
\examples{
\dontrun{
# return all lithology definitions
ex1 <- def_lithologies()
# return subsets of lithology definitions
ex2 <- def_lithologies(lithology = "sandstone")
ex3 <- def_lithologies(lithology_type = "sedimentary")
ex4 <- def_lithologies(lithology_id = c(1,5))
}
}
\author{
William Gearty
}
