% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_lithology_att.R
\name{def_lithology_att}
\alias{def_lithology_att}
\title{FUNCTION_TITLE}
\usage{
def_lithology_att(
  lithology_att = NULL,
  att_type = NULL,
  lithology_att_id = NULL
)
}
\arguments{
\item{lithology_att}{\code{character}. The name of the lithology attribute (e.g., "cross-bedded") to return a definition for.}

\item{att_type}{\code{character}. The type of lithology attribute (e.g., "sed structure") to return a definition for.}

\item{lithology_att_id}{\code{integer}. The unique identification number(s) of lithology attribute(s) to return a definition for.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\author{
AUTHOR \link{AUTHOR_2}
}
