% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_minerals.R
\name{def_minerals}
\alias{def_minerals}
\title{FUNCTION_TITLE}
\usage{
def_minerals(mineral = NULL, mineral_type = NULL, element = NULL)
}
\arguments{
\item{mineral}{\code{character}. The name of a mineral (e.g., "calcite") to return a definition for.}

\item{mineral_type}{\code{character}. The type of mineral (e.g., "feldspar") to return a definition for.}

\item{element}{\code{character}. The chemical element abbreviation of a mineral to return a definition for (e.g., "CaCO3"). Note: abbreviations are case sensitive.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\author{
AUTHOR \link{AUTHOR_2}
}
