% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_projects.R
\name{def_projects}
\alias{def_projects}
\title{Retrieve metadata for Macrostrat projects}
\usage{
def_projects(project_id = NULL)
}
\arguments{
\item{project_id}{\code{integer}. The unique identification number(s) of the
desired project(s) to return a definition for.}
}
\value{
A \code{dataframe} containing:
\describe{
\item{project_id}{The unique identifier of the Macrostrat project}
\item{project}{The name of the Macrostrat project}
\item{descrip}{Description of the Macrostrat project}
\item{timescale_id}{The unique identifier of the timescale used in the
Macrostrat project}
\item{t_cols}{The total number of Macrostrat columns associated with the
project}
\item{in_process_cols}{The total number of in-progress Macrostrat columns
associated with the project}
\item{obsolete_cols}{The total number of obsolete Macrostrat columns
associated with the project}
\item{t_units}{The total number of Macrostrat units associated with the
project}
\item{area}{The area covered by the Macrostrat project}
}
}
\description{
Obtain information concerning the different Macrostrat
projects. By default, data for all projects are returned.
}
\examples{
\dontrun{
if (interactive()) {
  # projects <- def_projects()
  # projects <- def_projects(project_id = 4)
}
}
}
\author{
Bethany Allen \link{AUTHOR_2}
}
