% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_references.R
\name{def_references}
\alias{def_references}
\title{Define references}
\usage{
def_references(ref_id = NULL)
}
\arguments{
\item{ref_id}{\code{integer}. The unique identification number(s) of the
desired reference(s) to return a definition for. If \code{NULL} (default), all
references are returned.}
}
\value{
A \code{data.frame} containing the following columns:
\itemize{
\item \code{ref_id}: The unique identification number(s) for the
reference(s).
\item \code{pub_year}: The year of publication of the reference.
\item \code{author}: The author(s) of the reference.
\item \code{ref}: The name of the reference.
\item \code{doi}: The digital object identifier of the reference.
\item \code{url}: A URL where additional information, the source, or
contributing publication can be found.
\item \code{t_units}: The total number of Macrostrat units associated with
the reference.
}
}
\description{
A function to retrieve the definitions for one or more published
scientific references in the Macrostrat database. By default, all reference
definitions are returned.
}
\section{Developer(s)}{

William Gearty
}

\section{Reviewer(s)}{

Bethany Allen
}

\examples{
\donttest{
# Return all references
ex1 <- def_references()
# Return a subset of references
ex2 <- def_references(ref_id = c(1, 2, 3))
}
}
\seealso{
Macrostrat database metadata:
\code{\link{def_projects}()},
\code{\link{def_sources}()},
\code{\link{get_stats}()}
}
\concept{meta}
