% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_sources.R
\name{def_sources}
\alias{def_sources}
\title{Retrieve geologic map source definitions}
\usage{
def_sources(
  source_id = NULL,
  lat = NULL,
  lng = NULL,
  shape = NULL,
  buffer = NULL,
  scale = NULL,
  sf = TRUE
)
}
\arguments{
\item{source_id}{\code{integer}. The unique identification number(s) of the
desired source(s) to return a definition for.}

\item{lat}{\code{numeric}. A valid latitude in decimal degrees to return a
source definition for. Must also supply \code{lng}.}

\item{lng}{\code{numeric}. A valid longitude in decimal degrees to return a
source definition for. Must also supply \code{lat}.}

\item{shape}{\code{character}. A valid well-known text (WKT) representation
of geometry, such as "POINT(30 10)" or "POLYGON((30 10, 40 40, 20 40, 10
20))", to return a source definition for.}

\item{buffer}{\code{integer}. The geographic buffer (in meters) that should
be applied to the specified \code{shape}.}

\item{scale}{\code{character}. The desired Burwell scale, either: "tiny"
(global), "small" (continental), "medium" (regional), or "large" (local).}

\item{sf}{\code{logical}. Should the results be returned as an \code{sf} object?
Defaults to \code{TRUE}. If \code{FALSE}, a \code{data.frame} is returned.}
}
\value{
A \code{data.frame} containing the following columns:
\itemize{
\item \code{source_id}: Identification number of the geologic map source.
\item \code{name}: Name of the geologic map source.
\item \code{url}: URL where additional information, the source, or
contributing publication can be found.
\item \code{ref_title}: Title of reference for geologic map source.
\item \code{authors}: Authors of geologic map source.
\item \code{ref_year}: Year of reference publication.
\item \code{ref_source}: Original publication source of the reference.
\item \code{isbn_doi}: The ISBN or DOI for the reference.
\item \code{scale}: The Macrostrat scale the geologic map source belongs
to.
\item \code{features}: The total number of features (i.e., outcrop shape
and point elements) associated with the geologic map source.
\item \code{area}: The total geographic area of the geologic map source
in km\ifelse{html}{\out{<sup>2</sup>}}{\eqn{^2}}.
}
If \code{sf} is \code{TRUE} (the default), an \code{sf} object is returned instead, with
the same columns plus a "geometry" column that contains the spatial data.
}
\description{
A function to retrieve the definitions for one or more geologic
map sources in the Macrostrat database. By default, all source definitions
are returned.
}
\section{Developer(s)}{

William Gearty
}

\section{Reviewer(s)}{

Lewis A. Jones
}

\examples{
\donttest{
# Get all sources
ex1 <- def_sources()
# Get subset of sources
ex2 <- def_sources(source_id = c(1,2,4))
ex3 <- def_sources(lat = 43.03, lng = -89.4, scale = "large")
# Use WKT representation
library(sf)
line <- st_linestring(x = matrix(c(-122.3438, 37,-89.3527, 43.0582),
                                 byrow = TRUE, ncol = 2))
ex4 <- def_sources(shape = st_as_text(line), buffer = 100)
}
}
\seealso{
Geologic maps:
\code{\link{get_map_legends}()},
\code{\link{get_map_outcrop}()},
\code{\link{get_map_points}()}

Macrostrat database metadata:
\code{\link{def_projects}()},
\code{\link{def_references}()},
\code{\link{get_stats}()}
}
\concept{maps}
\concept{meta}
