% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_strat_names.R
\name{def_strat_names}
\alias{def_strat_names}
\title{FUNCTION_TITLE}
\usage{
def_strat_names(
  strat_name = NULL,
  interval_name = NULL,
  rank = NULL,
  concept_id = NULL,
  ref_id = NULL,
  strat_name_id = NULL,
  rule = NULL
)
}
\arguments{
\item{strat_name}{\code{character}. The name(s) of the desired stratigraphic unit(s) to return a definition for.}

\item{interval_name}{\code{character}. The chronostratigraphic interval name for which to return stratigraphic name definitions with any temporal overlap (e.g., "Campanian"). This returns only \code{strat_names} with an age declared in Macrostrat.}

\item{rank}{\code{character}. The formal stratigraphic rank abbreviation of the desired unit(s) (i.e., "SGp", "Gp", "SubGp", "Fm", "Mbr", or "Bed") to return a definition for.}

\item{concept_id}{\code{integer}. The unique identification number(s) of stratigraphic concepts (grouped \code{strat_name}s which represent the same entity) to return a definition for.}

\item{ref_id}{\code{character}. The unique identification number(s) of the desired reference(s). Only stratigraphic names linked to the specified \code{ref_id} will be returned.}

\item{strat_name_id}{\code{integer}. The unique identification number(s) of the desired stratigraphic name to return a definition for.}

\item{rule}{\code{character}. The hierarchical rule for returning stratigraphic name definitions (either "all" or "down"). The "all" option will return the entire stratigraphic name hierarchy (parents and children of all levels) that the specified request belongs to. The "down" option will return any direct children of the requested stratigraphic name. Can only be used with \code{strat_name_id}.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\author{
AUTHOR \link{AUTHOR_2}
}
