% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_macrostrat.R
\name{fetch_macrostrat}
\alias{fetch_macrostrat}
\title{Fetch data from Macrostrat}
\usage{
fetch_macrostrat(path, query = NULL, format = "csv", ...)
}
\value{
User requested data, either a:
\item{data.frame}{}
\item{rast}{}
}
\description{
This function fetches data from \href{https://macrostrat.org}{Macrostrat}
via the Application Programming Interface (API). This is the core
function for handling user requests.
}
\examples{
path <- "defs/intervals?format=csv&timescale="
query <- "International Periods"
fetch_macrostrat(path = path, query = query, format = "csv")
}
