% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_args}
\alias{filter_args}
\title{Filters NULL and FALSE arguments from a list}
\usage{
filter_args(x)
}
\arguments{
\item{x}{\code{list}. The user-query list}
}
\value{
A \code{NULL} and \code{FALSE} filtered list of \code{x}.
}
\description{
This function checks whether any elements in a list are
NULL or FALSE and removes them.
}
\keyword{internal}
