% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_null}
\alias{filter_null}
\title{Filters NULL arguments from a list}
\usage{
filter_null(x)
}
\arguments{
\item{x}{\code{list}. The user-query list}
}
\value{
A \code{NULL} filtered list of \code{x}.
}
\description{
This function checks whether any elements in a list are
NULL and removes them.
}
