% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_age_model.R
\name{get_age_model}
\alias{get_age_model}
\title{Retrieve column age models}
\usage{
get_age_model(column_id = NULL, section_id = NULL)
}
\arguments{
\item{column_id}{\code{integer}. Filter age models to those associated with
column(s) as specified by their unique identification number(s).}

\item{section_id}{\code{integer}. Filter age models to those associated with
section(s) as specified by their unique identification number(s).}
}
\value{
A \code{dataframe} containing, for each boundary between units for
each retrieved age model:
\describe{
\item{boundary_id}{The unique identifier of the boundary between two units}
\item{col_id}{The unique identifier of the Macrostrat column}
\item{section_id}{The unique identifier of the Macrostrat section}
\item{interval_id}{The unique identifier of the time interval which the
boundary lies within}
\item{interval_name}{The name of the time interval which the boundary lies
within}
\item{age_bottom}{The numerical lower limit on the age of the boundary}
\item{age_top}{The numerical upper limit on the age of the boundary}
\item{rel_position}{The relative position of the boundary}
\item{model_age}{The age of the boundary, as correlated with other ages in
the column}
\item{boundary_status}{The method used to determine the age of the boundary}
\item{boundary_type}{The type of boundary}
\item{boundary_position}{The position of the boundary}
\item{unit_below}{The Macrostrat unit beneath the given boundary}
\item{unit_above}{The Macrostrat unit above the given boundary}
\item{ref_id}{The unique identifier of the reference}
}
}
\description{
Obtain information about the age models for columns. This
largely consists of the temporal boundaries applied to the column's
constituent units.
}
\details{
The Macrostrat age models are used to determine the ages of the
boundaries between units in an internally consistent manner, allowing each
boundary to be dated with respect to all age information available for the
column.
}
\examples{
\dontrun{
if (interactive()) {
  # age_model <- get_age_model(column_id = c(503, 504))
}
}
}
\author{
Bethany Allen \link{AUTHOR_2}
}
