% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_available_content}
\alias{get_available_content}
\title{Get available content}
\usage{
get_available_content(
  x = NULL,
  endpoint = NULL,
  query = NULL,
  var = NULL,
  available = TRUE
)
}
\arguments{
\item{x}{\code{character}. The requested content value (e.g. "international
ages").}

\item{endpoint}{\code{character}. The relevant Macrostrat API endpoint.}

\item{query}{\code{list}. The variable to extract from the returned
object to check \code{x} against.}

\item{var}{\code{character}. The variable to filter the content by.}

\item{available}{\code{logical}. Should all available content be returned
for the desired \code{query}?}
}
\value{
A \code{vector} of the available content or the matched \code{x}
value.
}
\description{
This function checks whether the requested content is available. It can be
used to either return all available content for the requested variable or
to standardise the requested content.
}
\details{
This function checks whether the requested content is available
via Macrostrat. It can also be used to return a vector of all available
content for the desired \code{query}.
}
