% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eodp.R
\name{get_eodp}
\alias{get_eodp}
\title{Retrieve core data from ocean drilling programs}
\usage{
get_eodp(column_id = NULL, site = NULL, leg = NULL, program = NULL, sf = FALSE)
}
\arguments{
\item{column_id}{\code{integer}. Filter cores by their unique identification
number(s).}

\item{site}{\code{character}. Filter cores to those from a particular
drilling site (e.g., "U1351").}

\item{leg}{\code{character}. Filter cores to those from a particular drilling
leg or expedition (e.g., "317").}

\item{program}{\code{character}. Filter cores to those from a particular
drilling program ("DSDP", "IODP", or "ODP").}

\item{sf}{\code{logical}. Should the results be returned as an \code{sf} object?
Defaults to \code{FALSE}.}
}
\value{
A \code{dataframe} containing, for each retrieved core:
\describe{
\item{col_group}{The name of the drilling program and leg/expedition}
\item{site_hole}{The name of the drilling site and hole}
\item{date_started}{The date on which drilling commenced}
\item{ref_id}{The unique identifier of the reference}
\item{col_id}{The unique identifier of the Macrostrat column}
\item{lat}{Decimal degree latitude of the core}
\item{lng}{Decimal degree longitude of the core}
\item{top_depth}{A vector describing the height of the top of each unit}
\item{bottom_depth}{A vector describing the height of the bottom of each
unit}
\item{primary_lith}{A vector giving the name of the primary lithology of
each unit}
\item{lith_id}{A vector giving the unique identifier of the primary
lithology of each unit}
\item{minor_lith}{A vector giving the name of the primary lithology of
each unit}
}
If sf = TRUE, an \code{sf} object is outputted instead.
}
\description{
Call data from the
\href{https://eodp.github.io}{Extending Ocean Drilling Pursuits (eODP)}
project, a collation of sedimentary description data
from ocean drilling cores. This includes cores from the Deep Sea Drilling
Project (DSDP), Integrated Ocean Drilling Program (IODP), and Ocean Drilling
Program (ODP).
}
\details{
More information can be found about the inputs for this function by
using \code{\link[=def_drilling_sites]{def_drilling_sites()}}.
}
\section{References}{

Sessa JA, Fraass AJ, LeVay LJ, Jamson KM, and Peters SE. (2023). The
Extending Ocean Drilling Pursuits (eODP) Project: Synthesizing Scientific
Ocean Drilling Data. Geochemistry, Geophysics, Geosystems, 24 (3)
e2022GC010655. \doi{10.1029/2022GC010655}.
}

\examples{
\dontrun{
if (interactive()) {
  # cores <- get_eodp(column_id = c(5081, 5082))
  # cores <- get_eodp(site = "U1351")
  # cores <- get_eodp(leg = "317")
  # cores <- get_eodp(program = "IODP")
}
}
}
\seealso{
External data:
\code{\link{def_drilling_sites}()},
\code{\link{def_measurements}()},
\code{\link{get_fossils}()},
\code{\link{get_measurements}()}
}
\author{
Bethany Allen
}
\concept{external}
