% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lithology.R
\name{get_lithology}
\alias{get_lithology}
\title{Get lithology data from Macrostrat}
\usage{
get_lithology(x = NULL, by = "name", available = FALSE)
}
\arguments{
\item{x}{\code{character}. The name of a lithology available via
\href{https://macrostrat.org/api/defs/lithologies?all}{Macrostrat}.}

\item{by}{\code{character}. The rank to search \code{name} by, should be either:
"name", "group", "type", or "class".}

\item{available}{\code{logical}. If set to \code{TRUE}, a vector of the
available lithologies will be returned. If set to \code{FALSE} (default),
the requested lithology will be checked against those available.}
}
\value{
A \code{vector} of available lithologies or a \code{data.frame} of
the requested
\href{https://macrostrat.org/api/defs/lithologies?all}{Macrostrat} lithology
containing at least the following columns:
\item{lith_id}{Lithology ID number.}
\item{name}{Name of lithology.}
\item{type}{Type of lithology.}
\item{group}{Group of lithology.}
\item{class}{Class of lithology.}
\item{color}{Assigned lithology colour.}
}
\description{
This function gets lithology data from \href{http://macrostrat.org}{Macrostrat}.
}
\details{
This function returns a \code{data.frame} of the requested
lithology and related data. Available
\href{https://macrostrat.org/api/defs/lithologies?all}{Macrostrat} lithologies
can be requested via the \code{available} argument. Note: if this argument is
set to \code{TRUE}, only a vector of available lithologies are returned.
To return the desired lithology, this argument should be set to
\code{FALSE}.
}
\examples{
# Get lithology by name
get_lithology(by = "name", available = TRUE)
# Get lithology by type
get_lithology(by = "type", available = TRUE)
# Get lithology by group
get_lithology(by = "group", available = TRUE)
# Get lithology by class
get_lithology(by = "class", available = TRUE)
# Get all sedimentary lithologies
get_lithology(x = "sedimentary", by = "class", available = FALSE)
# Get specific lithology
get_lithology(x = "limestone", by = "name", available = FALSE)
}
