% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_macrostrat.R
\name{GET_macrostrat}
\alias{GET_macrostrat}
\title{Fetch data from Macrostrat}
\usage{
GET_macrostrat(path, query = NULL, format = "csv", ...)
}
\value{
User requested data, either a \code{data.frame} or...
}
\description{
This function fetches data from \href{https://macrostrat.org}{Macrostrat}
via the Application Programming Interface (API). This is the core
function for handling user requests.
}
