% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measurements.R
\name{get_measurements}
\alias{get_measurements}
\title{Obtain geological measurements}
\usage{
get_measurements(
  measure_id = NULL,
  measurement_id = NULL,
  measurement = NULL,
  measurement_type = NULL,
  measurement_class = NULL,
  measuremeta_id = NULL,
  measure_phase = NULL,
  column_id = NULL,
  section_id = NULL,
  unit_id = NULL,
  interval_name = NULL,
  lithology_id = NULL,
  lithology_type = NULL,
  lithology_class = NULL,
  project_id = NULL,
  sf = FALSE
)
}
\arguments{
\item{measure_id}{\code{integer}. Filter measurements to those of one or more
unique identification number(s).}

\item{measurement_id}{\code{integer}. Filter measurements to those of one or
more measurement definition(s) as specified by their unique identification
number(s).}

\item{measurement}{\code{character}. Filter measurements to those of one or
more measurement definition(s) as specified by their name (e.g., "Pb" or
"transport direction").}

\item{measurement_type}{\code{character}. Filter measurements to those of one
or more measurement definition type(s) (e.g., "major elements",
"geochronological").}

\item{measurement_class}{\code{character}. Filter measurements to those of
one or more measurement definition class(es) (e.g., "geochemical",
"geophysical").}

\item{measuremeta_id}{\code{integer}. Filter measurements to those of one or
more measurement metadata(s) (usually refers to a sample) as specified by
their unique identification number(s).}

\item{measure_phase}{\code{character}. Filter measurements by the phase from
which the measurement was taken (e.g., "zircon", "whole rock").}

\item{column_id}{\code{integer}. Filter measurements to those within one or
more column(s) as specified by their unique identification number(s).}

\item{section_id}{\code{integer}. Filter measurements to those within one or
more section(s) as specified by their unique identification number(s).}

\item{unit_id}{\code{integer}. Filter measurements to those within one or
more unit(s) as specified by their unique identification number(s).}

\item{interval_name}{\code{character}. Filter measurements to those from
units within or touching a named interval(s) (e.g., "Permian").}

\item{lithology_id}{\code{integer}. Filter measurements to those of one or
more lithology(ies) as specified by their unique identification number(s).}

\item{lithology_type}{\code{character}. Filter measurements to those of one
or more lithology type(s) (e.g., "carbonate", "siliciclastic").}

\item{lithology_class}{\code{character}. Filter measurements to those of one
or more lithology class(es) (e.g., "sedimentary", "igneous",
"metamorphic").}

\item{project_id}{\code{integer}. Filter measurements by the unique
identification number(s) of the Macrostrat project(s).}

\item{sf}{\code{logical}. Should the results be returned as an \code{sf} object?
Defaults to \code{FALSE}.}
}
\value{
A \code{dataframe} containing, for each retrieved measurement:
\describe{
\item{measurement_id}{The unique identifier of the set of measurements}
\item{measuremeta_id}{The unique identifier (sample number) of the
measurement}
\item{measurement}{The name of the type of measurement obtained}
\item{measure_units}{The units of the measurement}
\item{measure_phase}{The phase from which the measurement was taken}
\item{method}{The method used to generate the measurement}
\item{n}{The number of observations or measurements}
\item{ref_id}{The unique identifier of the reference}
\item{sample_name}{Name of the sample for the measurement}
\item{geo_unit}{The Macrostrat unit from which the measurement was
taken}
\item{samp_lith}{A lithological description of the rock from which the
measurement was taken}
\item{samp_lith_id}{The unique identifier of the lithological description of
the rock from which the measurement was taken}
\item{samp_desc}{A description of the sample used to generate the
measurement}
\item{samp_age}{The geological time interval assigned to the measurement}
\item{lat}{Decimal degree latitude of the measurement}
\item{lng}{Decimal degree longitude of the measurement}
\item{unit_id}{The unique identifier of the Macrostrat unit from which the
measurement was taken}
\item{unit_rel_pos}{The relative positive of the sample or measurement
within the unit}
\item{col_id}{The unique identifier of the Macrostrat column from which the
measurement was taken}
\item{strat_name_id}{The unique identifier of the stratigraphic name}
\item{match_basis}{A terse descriptor of how the measuremeta data was linked
to the Macrostrat unit}
\item{ref}{The name of the reference}
\item{measure_value}{The value of the measurement}
\item{measure_error}{The reported error on the measurement value}
\item{measure_position}{The position of the measurement in the Macrostrat
column}
\item{measure_n}{The number of measurements used to generate the
measure_value; if greater than one, usually used to produce the
measure_error}
\item{sample_no}{The sample number for the measurement}
\item{error_units}{The units of the error}
}
If sf = TRUE, an \code{sf} object is outputted instead.
}
\description{
Retrieve a range of measurements relevant to making geological
inferences.
}
\details{
More information can be found about the inputs and outputs for this
function by using \code{\link[=def_measurements]{def_measurements()}}.
}
\examples{
\dontrun{
if (interactive()) {
  # age_model <- get_measurements(measure_id = c(353, 354))
  # age_model <- get_measurements(measurement_id = 42)
  # age_model <- get_measurements(column_id = 84, measurement = "SiO2")
}
}
}
\author{
Bethany Allen \link{AUTHOR_2}
}
