% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sections.R
\name{get_sections}
\alias{get_sections}
\title{FUNCTION_TITLE}
\usage{
get_sections(
  col_id,
  unit_id,
  section_id,
  interval_name,
  int_id,
  age,
  age_top,
  age_bottom,
  lith,
  lith_id,
  lith_group,
  lith_type,
  lith_class,
  lith_att,
  lith_att_id,
  lith_att_type,
  environ,
  environ_id,
  environ_type,
  environ_class,
  econ,
  econ_id,
  econ_type,
  econ_class,
  cltn_id,
  strat_name,
  strat_name_id,
  lat,
  lng,
  adjacents,
  project_id,
  response = NULL
)
}
\arguments{
\item{col_id}{\code{integer}. The unique identification number(s) of the columns(s) to return.}

\item{unit_id}{\code{integer}. The unique identification number(s) of the unit(s) to filter the results by.}

\item{section_id}{\code{integer}. The unique identification number(s) of the section(s) to filter the results by.}

\item{interval_name}{\code{character}. The name of a chronostratigraphic time interval to filter the results by.}

\item{int_id}{\code{integer}. The unique identification number(s) of a chronostratigraphic time interval to filter the results by.}

\item{age}{\code{numeric}. Filter sections to those that overlap this numerical age in millions of years before present.}

\item{age_top}{\code{numeric}. Filter sections to those that overlap with the age range between this and age_bottom. Should be in millions of years before present and less than age_bottom.}

\item{age_bottom}{\code{numeric}. Filter sections to those that overlap with the age range between this and age_top. Should be in millions of years before present and greater than age_top.}

\item{lith}{\code{character}. Filter sections to those containing a named lithology (e.g., shale, sandstone).}

\item{lith_id}{\code{integer}. Filter sections to those containing a lithology identified by its unique identification number.}

\item{lith_group}{\code{character}. Filter sections to those containing a named lithology group (e.g., sandstones, mudrocks, unconsolidated)}

\item{lith_type}{\code{character}. Filter sections to those containing a named lithology type (e.g., carbonate, siliciclastic).}

\item{lith_class}{\code{character}. Filter sections to those containing a named general lithology (sedimentary, igneous, metamorphic).}

\item{lith_att}{\code{character}. Filter sections to those containing a named lithology attribute (e.g. fine, olivine, poorly washed).}

\item{lith_att_id}{\code{integer}. Filter sections to those containing a lithology attribute identified by its unique identification number.}

\item{lith_att_type}{\code{character}. Filter sections to those containing a named category of lithology attribute (e.g. grains, lithology, bedform).}

\item{environ}{\code{character}. Filter sections to those containing a named environment.}

\item{environ_id}{\code{integer}. Filter sections to those containing an environment identified by its unique identification number.}

\item{environ_type}{\code{character}. Filter sections to those containing a named group of environments.}

\item{environ_class}{\code{character}. Filter sections to those containing a named general environment.}

\item{econ}{\code{character}. Filter sections to those containing a named economic attribute.}

\item{econ_id}{\code{integer}. Filter sections to those containing an economic attribute identified by its unique identification number.}

\item{econ_type}{\code{character}. Filter sections to those containing a named economic attribute type.}

\item{econ_class}{\code{character}. Filter sections to those containing a named economic attribute class.}

\item{cltn_id}{\code{integer}. Filter sections to those containing one or more Paleobiology Database collections identified by their unique identification numbers.}

\item{strat_name}{\code{character}. Filter sections to those containing a unit that fuzzy matches a stratigraphic name.}

\item{strat_name_id}{\code{integer}. Filter sections by a single or comma-separated list of stratigraphic IDs.}

\item{lat}{\code{numeric}. Return the section at a particular decimal degree latitude (WGS84). Must also specify \code{lng}.}

\item{lng}{\code{numeric}. Return the section at a particular decimal degree longitude (WGS84). Must also specify \code{lat}.}

\item{adjacents}{\code{logical}. If lat/lng or col_id is specified, should all sections that touch the specified column be returned?}

\item{project_id}{\code{integer}. Filter sections to those contained within a Macrostrat project identified by its unique identification number.}

\item{response}{\code{character}. How much metadata should be returned? Options (in increasing amount of metadata) are "short" and "long"., Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
AUTHOR \link{AUTHOR_2}
}
