% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_stats}
\alias{get_stats}
\title{Get Macrostrat project statistics from Macrostrat}
\usage{
get_stats(x = NULL, available = FALSE)
}
\arguments{
\item{x}{\code{character}. The name of a project available via
\href{https://macrostrat.org/api/stats?all}{Macrostrat}.}

\item{available}{\code{logical}. If set to \code{TRUE}, a vector of the
available projects will be returned. If set to \code{FALSE} (default),
the requested project will be checked against those available.}
}
\value{
A \code{vector} of available projects or a \code{data.frame} of
the requested
\href{https://macrostrat.org/api/stats?all}{Macrostrat} project statistics
containing:
\item{project_id}{The project ID number.}
\item{project}{The name of the project.}
\item{columns}{The number of geological columns associated with the
project.}
\item{packages}{The number of geological packages associated with the
project.}
\item{units}{The number of geological units associated with the project.}
\item{pbdb_collections}{The number of Paleobiology Database collections
associated with the project.}
\item{measurements}{The number of measurements associated with the
project.}
\item{t_polys}{The number of polygons(?) associated with the project.}
}
\description{
This function gets specific statistics on Macrostrat projects from
\href{http://macrostrat.org}{Macrostrat}.
}
\details{
This function returns a \code{data.frame} of the requested
project stats. Available
\href{https://macrostrat.org/api/stats?all}{Macrostrat} project statistics
can be requested via the \code{available} argument. Note: if this argument is
set to \code{TRUE}, only a vector of available project names are returned.
To return the desired project statistics, this argument should be set to
\code{FALSE}.
}
\examples{
# Get available timescales
available <- get_stats(available = TRUE)
# Get all stats
all <- get_stats()
# Get project specific
nz <- get_stats(x = "New Zealand")
}
