% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timescale.R
\name{get_timescale}
\alias{get_timescale}
\title{Get timescale data from Macrostrat}
\usage{
get_timescale(scale = "international periods")
}
\arguments{
\item{scale}{\code{character}. The name of a timescale available via
\href{https://macrostrat.org/api/defs/timescales?all}{Macrostrat}.}
}
\value{
A \code{data.frame} of the requested
\href{https://macrostrat.org/api/defs/timescales?all}{Macrostrat} timescale.
}
\description{
This function gets timescale data from \href{http://macrostrat.org}{Macrostrat}.
}
\details{
This function returns a \code{data.frame} of a desired timescale
available via
\href{https://macrostrat.org/api/defs/timescales?all}{Macrostrat}.
}
\examples{
timescale <- get_timescale(scale = "international age")
}
