% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrostrat_intervals.R
\name{macrostrat_intervals}
\alias{macrostrat_intervals}
\title{Retrieve Macrostrat hosted interval data}
\usage{
macrostrat_intervals(
  timescale = NULL,
  name = NULL,
  age = NULL,
  t_age = NULL,
  b_age = NULL,
  rule = NULL,
  all = TRUE,
  true_colors = NULL,
  int_id = NULL,
  timescale_id = NULL
)
}
\arguments{
\item{timescale}{\code{character}. The name of the desired timescale.
Available timescales can be retrieved via \link{macrostrat_timescales}.}

\item{name}{\code{character}. The name of the desired interval. If set to
NULL (default), all interval data for the respective timescale is
returned.}

\item{age}{\code{numeric}. Age in millions of years before present. All
intervals overlapping with this age will be returned.}

\item{t_age}{\code{numeric}. The minimum age bound interval data should be
returned for in millions of years before present. If set to NULL
(default), all interval data for the respective timescale is returned.}

\item{b_age}{\code{numeric}. The maximum age bound interval data should be
returned for in millions of years before present. If set to NULL
(default), all interval data for the respective timescale is returned.}

\item{rule}{\code{character}. How should intervals be returned for the
given \code{t_age} and \code{b_age}? Use "contains" to return all intervals that
fall entirely within \code{t_age} and \code{b_age}. Use "exact" to return all
intervals within boundaries equal to \code{t_age} and \code{b_age}. Use "loose" to
return all intervals touching the range of \code{t_age} and \code{b_age}. If
\code{t_age} and \code{b_age} are not provided, this argument is ignored.}

\item{all}{\code{logical}. Should all available interval definitions be
returned? Defaults to TRUE.}

\item{true_colors}{\code{logical}. Should the original international time
scale colors be returned? Defaults to TRUE.}

\item{int_id}{\code{integer}. The identification number of the desired
interval.}

\item{timescale_id}{\code{integer}. The identification number of the
desired timescale. Available timescale identification numbers can be
retrieved via \link{macrostrat_timescales}.}
}
\value{
A \code{data.frame} of time scales hosted by
\href{https://macrostrat.org}{Macrostrat} containing:
\itemize{
\item{int_id: }{The unique identification number of the interval.}
\item{name: }{The name of the interval.}
\item{abbrev: }{If available, the interval abbreviation.}
\item{t_age: }{The minimum age of the interval.}
\item{b_age: }{The maximum age of the interval.}
\item{int_type: }{The type (or rank) of interval (e.g. age, period).}
\item{timescales: }{The time scale the interval belongs to.}
\item{color: }{International time scale colors.}
}
}
\description{
This function fetches Macrostrat interval data based on
user-defined arguments via the \href{https://macrostrat.org/#api}{Macrostrat API}.
}
\details{
DETAILS
}
\examples{
# Get international ages
ages <- macrostrat_intervals(timescale = "international ages")
# Get international periods
periods <- macrostrat_intervals(timescale = "international periods")
# Get NALMAs
NALMAs <- macrostrat_intervals(timescale = "North American Land Mammal Ages")
# Get specific interval by name
Hettangian <- macrostrat_intervals(name = "Hettangian")
# Get specific interval by ID
Chattian <- macrostrat_intervals(int_id = 20L)
# Get international age by age
Gzhelian <- macrostrat_intervals(timescale = "international ages", age = 300)
# Get international age by age range
intervals <- macrostrat_intervals(timescale = "international ages", t_age = 0, b_age = 100)
}
\author{
Lewis A. Jones
}
