% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macrostrat_timescales.R
\name{macrostrat_timescales}
\alias{macrostrat_timescales}
\title{Retrieve Macrostrat hosted time scales}
\usage{
macrostrat_timescales()
}
\value{
A \code{data.frame} of time scales hosted by
\href{https://macrostrat.org}{Macrostrat} containing:
\itemize{
\item{timescale_id: }{The time scale identification number.}
\item{timescale: }{The name of the time scale.}
\item{n_intervals: }{The number of intervals within the time scale.}
\item{max_age: }{The maximum age of the intervals in the time scale.}
\item{min_age: }{The minimum age of the intervals in the time scale.}
\item{ref_id: }{The identification number of the reference the time scale
is based on.}
}
}
\description{
This function fetches all available time scales from
\href{https://macrostrat.org}{Macrostrat} via the \href{https://macrostrat.org/#api}{Macrostrat API}.
}
\examples{
# Get available time scales
macrostrat_timescales()
}
\author{
Lewis A. Jones
}
