% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmacrostrat-package.R
\docType{package}
\name{rmacrostrat-package}
\alias{rmacrostrat}
\alias{rmacrostrat-package}
\title{rmacrostrat: Fetch Geologic Data from the 'Macrostrat' Platform}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Work with the 'Macrostrat' (\url{https://macrostrat.org/}) Web Service (v.2, \url{https://macrostrat.org/api/v2}) to fetch geological data relevant to the spatial and temporal distribution of sedimentary, igneous, and metamorphic rocks as well as data extracted from them.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rmacrostrat.palaeoverse.org}
  \item \url{https://palaeoverse.org}
  \item \url{https://macrostrat.org}
  \item \url{https://github.com/UW-Macrostrat/macrostrat/issues}
  \item Report bugs at \url{https://github.com/palaeoverse/rmacrostrat/issues}
}

}
\author{
\strong{Maintainer}: Lewis A. Jones \email{LewisA.Jones@outlook.com} (\href{https://orcid.org/0000-0003-3902-8986}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item William Gearty \email{willgearty@gmail.com} (\href{https://orcid.org/0000-0003-0076-3262}{ORCID})
  \item Christopher D. Dean \email{christopherdaviddean@gmail.com} (\href{https://orcid.org/0000-0001-6471-6903}{ORCID})
  \item Bethany Allen \email{bethany.allen@bsse.ethz.ch} (\href{https://orcid.org/0000-0003-0282-6407}{ORCID})
}

}
\keyword{internal}
