% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_drilling_sites.R
\name{def_drilling_sites}
\alias{def_drilling_sites}
\title{Retrieve metadata associated with eODP}
\usage{
def_drilling_sites(program = NULL, leg = NULL, site = NULL, sf = NULL)
}
\arguments{
\item{program}{\code{character}. The name of a drilling program (i.e.,
"DSDP", "ODP", or "IODP") to return a definition for.}

\item{leg}{\code{character}. The unique identification number(s) of drilling
leg(s) to return a definition for.}

\item{site}{\code{character}. The unique identification number(s) of drilling
site(s) to return a definition for.}

\item{sf}{\code{logical}. Should the results be returned as an \code{sf} object
(defaults to \code{TRUE})? If \code{FALSE}, a \code{data.frame} is returned.}
}
\value{
An \code{sf} object containing, for each retrieved core:
\describe{
\item{epoch}{The name of the drilling program}
\item{leg}{The name of the leg/expedition}
\item{site}{The name of the drilling site}
\item{hole}{The name of the drilling hole}
\item{lat}{Decimal degree latitude of the core}
\item{lng}{Decimal degree longitude of the core}
\item{col_id}{The unique identifier of the Macrostrat column}
\item{col_group_id}{The unique identifier of the group to which the
Macrostrat column belongs}
\item{penetration}{The depth of the hole}
\item{cored}{}
\item{recovered}{}
\item{recovery}{}
\item{drilled_interval}{}
\item{drilled_intervals}{}
\item{cores}{}
\item{date_started}{The date on which drilling commenced}
\item{comments}{Written notes assigned to the core}
\item{ref_id}{The unique identifier of the reference}
}
If sf = FALSE, a \code{dataframe} is outputted instead.
}
\description{
Obtain metadata for variables associated with the
\href{https://eodp.github.io}{Extending Ocean Drilling Pursuits (eODP)} project.
By default, data for all drilling sites are returned.
}
\section{References}{

Sessa JA, Fraass AJ, LeVay LJ, Jamson KM, and Peters SE. (2023). The
Extending Ocean Drilling Pursuits (eODP) Project: Synthesizing Scientific
Ocean Drilling Data. Geochemistry, Geophysics, Geosystems, 24 (3)
e2022GC010655. \doi{10.1029/2022GC010655}.
}

\examples{
\dontrun{
if (interactive()) {
  # core_info <- def_drilling_sites(site = "U1547")
  # core_info <- def_drilling_sites(leg = "385")
}
}
}
\author{
Bethany Allen \link{AUTHOR_2}
}
