% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_timescales.R
\name{def_timescales}
\alias{def_timescales}
\title{Define timescales}
\usage{
def_timescales()
}
\value{
A \code{data.frame} containing the following columns:
\itemize{
\item \code{timescale_id}: The unique identification number of the
timescale.
\item \code{timescale}: The name of the timescale.
\item \code{n_intervals}: The number of intervals within the timescale.
\item \code{max_age}: The maximum age coverage of the timescale in
millions of years before present.
\item \code{min_age}: The minimum age coverage of the timescale in
millions of years before present.
\item \code{ref_id}: The unique identification number of the associated
reference.
}
}
\description{
A function to retrieve all timescale definitions from the
Macrostrat database. Function is called without user-specified arguments.
}
\section{Developer(s)}{

Lewis A. Jones
}

\section{Reviewer(s)}{

Christopher D. Dean
}

\examples{
\donttest{
 # Retrieve all timescale definitions
 ex1 <- def_timescales()
}
}
\seealso{
Timescales and time intervals:
\code{\link{def_intervals}()}
}
\concept{defs_time}
