% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{GET_macrostrat}
\alias{GET_macrostrat}
\title{Fetch data from Macrostrat}
\usage{
GET_macrostrat(path, query = list())
}
\arguments{
\item{path}{\code{character}. The API path to retrieve data from
Macrostrat.}

\item{query}{\code{character}. The API query to retreive data from
Macrostrat.}
}
\value{
A \code{data.frame} of user-requested data.
}
\description{
This function fetches data from \href{https://macrostrat.org}{Macrostrat}
via the Application Programming Interface (API). This is the core
function for handling user requests.
}
