% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_options}
\alias{get_options}
\title{Check available options}
\usage{
get_options(x = NULL, endpoint = NULL)
}
\arguments{
\item{x}{\code{character}. Which options should be checked? Choose from
"parameters", "output_formats", "examples", or "fields". If this argument
is set to \code{NULL} (default), all options are returned.}

\item{endpoint}{\code{character}. The endpoint that should be checked
such as "defs/timescales" or "columns".}
}
\value{
A \code{list} of available "parameters", "output_formats",
"examples", or "fields". If \code{x} and \code{endpoint} are set to \code{NULL},
the base routes are returned.
}
\description{
This function checks which options are available.
}
