% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tiles.R
\name{get_tiles}
\alias{get_tiles}
\title{Retrieve Macrostrat vectorized tiles}
\usage{
get_tiles(zoom = 0, x = NULL, y = NULL, scale = "carto", combined = TRUE)
}
\arguments{
\item{zoom}{\code{integer}. The unique identification number(s) of the map
outcrop shape element(s) to return.}

\item{x}{\code{integer}. The x index/indices of the tile(s) to retrieve. If
\code{NULL} (the default), tiles of all x indices will be retrieved and
combined.}

\item{y}{\code{integer}. The y index/indices of the tile(s) to retrieve. If
\code{NULL} (the default), tiles of all y indices will be retrieved and
combined.}

\item{scale}{\code{character}. The Macrostrat map scale to be used. Options
are "tiny" (global), "small" (continental), "medium" (regional), "large",
(local), or "carto" (blended). The "carto" scale (the default) is for
visualization purposes, and makes many assumptions about the relative
priority of each map. The layers and scales are seamlessly blended so no
scale-dependent decisions can or should be made.}

\item{combined}{\code{logical}. Whether all tiles should be combined. If
\code{TRUE} (the default), the function returns a single list with all
tiles combined (as described below). If \code{FALSE}, the function returns
a list of lists, each element of which represents the data for a single
tile.}
}
\value{
A \code{list} of length two containing the following elements:
\itemize{
\item \code{lines}: An \code{sf} object containing the geologic line
features such as faults, anticlines, and moraines. It includes 7
variables:
\itemize{
\item \code{map_id}: The unique identification number of the geologic
}
\item \code{units}: An \code{sf} object containing the geologic map
polygons. It includes 23 variables:
\itemize{
\item \code{map_id}: The unique identification number of the geologic
}
}
}
\description{
A function to retrieve map tile data from the
\href{https://tiles.macrostrat.org/}{Macrostrat tile server}. This includes
geologic line features such as faults, anticlines, and moraines, as well as
geologic map polygons. The function retrieves the data in vectorized tile
format.
}
\details{
Note that the tile indices (\code{x} and \code{y}) are zero-indexed,
meaning that the first tile in each dimension is \code{0}. The zoom level
is also zero-indexed, meaning that the first zoom level is \code{0}. For a
given zoom level, the possible tile indices are \code{0} to \code{2 ^ zoom
  - 1}.
}
\section{Developer(s)}{

William Gearty
}

\section{Reviewer(s)}{

Lewis A. Jones
}

\examples{
\donttest{
# get single carto tile of the world
ex1 <- get_tiles()
# get specific "large" tile
ex2 <- get_tiles(zoom = 6, x = 5, y = 5, scale = "tiny")
}
}
\seealso{
Geologic maps:
\code{\link{def_sources}()},
\code{\link{get_map_legends}()},
\code{\link{get_map_outcrop}()},
\code{\link{get_map_points}()}
}
\concept{maps}
