% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmacrostrat-package.R
\docType{package}
\name{rmacrostrat-package}
\alias{rmacrostrat}
\alias{rmacrostrat-package}
\title{rmacrostrat: Fetch Data from the Macrostrat Platform}
\description{
Work with the Macrostrat Web Service (\url{https://macrostrat.org/}) to fetch geological data relevant to the spatial and temporal distribution of sedimentary, igneous, and metamorphic rocks as well as data extracted from them.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rmacrostrat.palaeoverse.org}
  \item \url{https://palaeoverse.org}
  \item \url{https://macrostrat.org}
  \item \url{https://palaeoverse-community.github.io/rmacrostrat/}
  \item Report bugs at \url{https://github.com/palaeoverse-community/rmacrostrat/issues}
}

}
\author{
\strong{Maintainer}: Lewis A. Jones \email{LewisA.Jones@outlook.com} (\href{https://orcid.org/0000-0003-3902-8986}{ORCID})

Authors:
\itemize{
  \item William Gearty \email{willgearty@gmail.com} (\href{https://orcid.org/0000-0003-0076-3262}{ORCID})
}

}
\keyword{internal}
